% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cells.R
\name{annotate_hierarchy}
\alias{annotate_hierarchy}
\title{Make hierarchical named groups from SingleR annotation}
\usage{
annotate_hierarchy(group, reference)
}
\arguments{
\item{group}{label.fine annotation from SingleR}

\item{reference}{reference character string matching name of dataset to compare against (uses a selection from celldex package; default "MonacoImmuneData" covers most populations in PBMC). Alternatively, a custom SummarizedExperiment can be provided.}
}
\value{
list of hierarical cell type assignments
}
\description{
Only "MonacoImmuneData" supported at the moment.
}
