% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cells.R
\name{annotate_hierarchy_MonacoImmuneData}
\alias{annotate_hierarchy_MonacoImmuneData}
\title{Make hierarchical named groups from MonacoImmuneData annotation}
\usage{
annotate_hierarchy_MonacoImmuneData(group)
}
\arguments{
\item{group}{MonacoImmuneData$label.fine annotation}
}
\description{
A bit of a hacky approach for primarily for internal use.
}
