% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clusters.R
\name{annotate_merge_by_cluster}
\alias{annotate_merge_by_cluster}
\title{Merge annotation by cluster}
\usage{
annotate_merge_by_cluster(
  annotation,
  cluster,
  threshold = "max",
  collapse = "/"
)
}
\description{
Merges individual cell annotations by cluster using majority rule (the entire cluster is annotated as the individual annotation represented by most cells within the cluster)
}
