% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cells.R
\name{annotate_SingleR}
\alias{annotate_SingleR}
\title{annotate individual cells with SingleR}
\usage{
annotate_SingleR(
  data,
  reference = "MonacoImmuneData",
  labels_column = "label.fine",
  ...
)
}
\arguments{
\item{data}{matrix of normalized expression values}

\item{reference}{reference character string matching name of dataset to compare against (uses a selection from celldex package; default "MonacoImmuneData" covers most populations in PBMC). Alternatively, a custom SummarizedExperiment can be provided.}

\item{labels_column}{name of labels column in the reference dataset to be used (labels.main or labels.fine if using celldex reference)}

\item{...}{Passed on to SingleR::SingleR}
}
\value{
data.frame containing labels for each cell
}
\description{
annotate individual cells with SingleR
}
