% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{biexp_trans}
\alias{biexp_trans}
\title{Biexponential transformation (inspired by flowJo)}
\usage{
biexp_trans(lim = 5, decade.size = lim)
}
\description{
Biexponential transformation (inspired by flowJo)
}
