% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utility_functions.R
\name{colors_get_distinct}
\alias{colors_get_distinct}
\title{Get a vector of more or less distinct colors}
\usage{
colors_get_distinct()
}
\value{
vector of colors
}
\description{
Could be improved my manual curation.
}
