% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred_encircle.R
\name{dimred_encircle}
\alias{dimred_encircle}
\title{Encircle using KDE contours}
\usage{
dimred_encircle(
  x,
  y,
  ...,
  contour_level = "3\%",
  contour_factor = 1,
  bandwidth = c(10, 10)
)
}
\arguments{
\item{...}{Passed on toe ks:kde}

\item{contour_level}{Character string for which contour level to use (Default: "3%")}

\item{contour_factor}{contour_levels only available for whole percentages. If an intermediate level or a level below "1%" is needed, the contour level value is divided by this factor.}

\item{bandwidth}{Bandwidth passeds on the ks::kde}

\item{data}{matrix or data frame with x and y values as the first two columns}
}
\value{
data.frame of contour points
}
\description{
Calculate kernel density estimates and extract contour lines at a given level to make flexible encircling of points.
}
\details{
Modified from: https://stackoverflow.com/questions/23437000/how-to-plot-a-contour-line-showing-where-95-of-values-fall-within-in-r-and-in
}
