% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred_encircle.R
\name{dimred_encircle_add}
\alias{dimred_encircle_add}
\title{Add encircling to ggplot}
\usage{
dimred_encircle_add(
  data,
  dim1 = 1,
  dim2 = 2,
  color_by = NULL,
  colors = NULL,
  size = 1,
  alpha = 0,
  line_alpha = 0.5,
  linetype = "solid",
  show_legend = FALSE,
  rescale = TRUE,
  borderline = FALSE,
  borderline_color = alpha("black", 0.5),
  ...
)
}
\arguments{
\item{data}{matrix or data.frame containing dimensions and grouping parameters. All columns that are not given as dim1 or dim2 (see below), will be used in grouping the resulting KDE contour.}

\item{dim1}{column position or column name of the first dimension in `data`}

\item{dim2}{column position or column name of the second dimension in `data`}

\item{color_by}{column position or column name for coloring the geom_path}

\item{size}{Line size for geom_polygon}

\item{alpha}{Alpha for geom_polygon}

\item{line_alpha}{Alpha for line of geom_polygon}

\item{linetype}{Line type for geom_polygon}

\item{show_legend}{Should legend be drawn?}

\item{rescale}{Should scales be adjusted to make sure contours are complete? This will override previous scale_x_ and scale_y_.}

\item{borderline}{Should line have a border?}

\item{borderline_color}{If borderline=TRUE, what color should be used for border?}

\item{...}{Passed on to scutility::dimred_encircle_groups()}
}
\value{
list of ggplot elements
}
\description{
Encircle by Kernel Density Estimate contours
}
