% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred_encircle.R
\name{dimred_encircle_groups}
\alias{dimred_encircle_groups}
\title{Encircle groups using KDE contours}
\usage{
dimred_encircle_groups(data, dim1 = 1, dim2 = 2, ...)
}
\arguments{
\item{data}{matrix or data.frame containing dimensions and grouping parameters. All columns that are not given as dim1 or dim2 (see below), will be used in grouping the resulting KDE contour.}

\item{dim1}{column position or column name of the first dimension in `data`}

\item{dim2}{column position or column name of the second dimension in `data`}

\item{...}{Passed on to scutility::dimred_encircle() - CURRENTLY NOT IMPLEMENTED}
}
\value{
data.frame of contour points and grouping variables
}
\description{
For each groupings, calculate kernel density estimates and extract contour lines at a given level to make flexible encircling of points.
}
