% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_labels.R
\name{dimred_labels_add}
\alias{dimred_labels_add}
\title{Add labels to DimPlot (or similar ggplot2)}
\usage{
dimred_labels_add(
  labels,
  colors = NULL,
  label_by = 1,
  color_by = 1,
  dim1 = 2,
  dim2 = 3,
  seed = 112,
  label.size = NA,
  fontface = "bold",
  label.padding = unit(0.01, units = "npc"),
  na.rm = TRUE,
  fill = alpha(c("white"), 0.75),
  ...
)
}
\arguments{
\item{labels}{data.frame of labels and their positions (i.e. created by scutility::seurat_dimred_cluster_labels)}

\item{colors}{named vector of colors for each label}

\item{label_by}{name or index of the column in labels that should be used for labelling}

\item{color_by}{name or index of the column in labels that should be used for coloring}

\item{dim1}{name or index of the column in labels that should be used for x}

\item{dim2}{name or index of the column in labels that should be used for y}

\item{...}{Passed on to ggrepel::geom_label_repel}
}
\value{
list of ggplot2 functions to be added to an existing plot
}
\description{
Sets a new color scale for the added annotation and sets up default aesthetics
}
