% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_functions.R
\name{droplet_barcodes_debris}
\alias{droplet_barcodes_debris}
\title{Get debris barcodes based on rank cutoff and called cell barcodes}
\usage{
droplet_barcodes_debris(barcodes_cells, debris_cutoff, ranking_matrix)
}
\arguments{
\item{barcodes_cells}{Barcodes for droplet called as cell-containing}

\item{debris_cutoff}{Cutoff for which droplets are likely to contain cells or debris (can be set to estimated number of cells captured)}

\item{ranking_matrix}{Count matrix used for ranking droplets}
}
\value{
vector of barcodes
}
\description{
Get debris barcodes based on rank cutoff and called cell barcodes
}
