% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_functions.R
\name{droplet_counts_per_compartment}
\alias{droplet_counts_per_compartment}
\title{Droplet counts per compartment}
\usage{
droplet_counts_per_compartment(
  matrix,
  compartment,
  sum_function = sum,
  remaining_include = TRUE,
  remaining_name = "empty"
)
}
\arguments{
\item{matrix}{Count matrix}

\item{compartment}{A named list of compartments and their barcodes (i.e. list("cells"=c("BC1", "BC2, ...), "debris"=c("BC101", "BC222, ...)))}

\item{sum_function}{Function applied to each group of barcodes}

\item{remaining_include}{Should barcodes that are not included in compartment list be included as its own compartment (i.e. for empty droplets)}

\item{remaining_name}{If remaining barcodes are included as a compartment, what should it be called?}
}
\value{
Vector or matrix depending on the sum function
}
\description{
Calculate count sums for subsets of columns in a matrix
}
