% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_load.R
\name{modalities_load_cellranger_count}
\alias{modalities_load_cellranger_count}
\title{Load Cell Ranger data and divide by modality}
\usage{
modalities_load_cellranger_count(
  path,
  modalities = c("RNA", "ADT", "HTO"),
  folder = "outs/raw_feature_bc_matrix",
  hto.pattern = "^hto",
  gex.listname = "Gene Expression",
  adt.listname = "Antibody Capture"
)
}
\arguments{
\item{path}{Path to cellranger output (outs) folder}

\item{modalities}{Vector of modalities to be extracted from Cell Ranger output}

\item{folder}{Folder containing count matrix (raw_feature_bc_matrix or filtered_feature_bc_matrix)}

\item{hto.pattern}{Pattern in feature name distinguishing hastag (HTOs) from other antibody derived tags (ADTs)}

\item{gex.listname}{Name of dataframe containing gene expression counts (after Seurat::Read10X)}

\item{adt.listname}{Name of dataframe containing ADT counts (after Seurat::Read10X)}
}
\value{
list of modality count matrices
}
\description{
Load Cell Ranger data and divide by modality
}
