% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_load.R
\name{modalities_load_kallisto}
\alias{modalities_load_kallisto}
\title{Load Kallisto data by modality}
\usage{
modalities_load_kallisto(
  paths,
  modalities = c("ADT", "HTO"),
  barcode_suffix = "",
  folder = "counts_unfiltered",
  ...
)
}
\arguments{
\item{paths}{Paths to kallisto output folders}

\item{modalities}{Vector of modalities to be loaded from kallisto output}

\item{barcode_suffix}{Suffix added to the end of each cell barcode name (to make it compatible with Cell Ranger output)}

\item{folder}{Folder containing count matrix}
}
\value{
list of modality count matrices
}
\description{
Load Kallisto data by modality
}
