% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_balance.R
\name{mtx_load_long}
\alias{mtx_load_long}
\title{Load MTX file as long format data.frame}
\usage{
mtx_load_long(
  path,
  file_matrix = "matrix.mtx",
  file_barcodes = "barcodes.tsv",
  file_features = "genes.tsv",
  matrix_sep = " ",
  matrix_skip = 3,
  matrix_colnames = c("feature", "barcode", "count")
)
}
\description{
Load MTX file as long format data.frame
}
