% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_balance.R
\name{mtx_long_save}
\alias{mtx_long_save}
\title{Save long format mtx data.frame as mtx}
\usage{
mtx_long_save(
  mtx,
  path,
  file_matrix = "matrix.mtx",
  file_barcodes = "barcodes.tsv",
  file_features = "genes.tsv",
  matrix_header = c("\%\%MatrixMarket matrix coordinate integer general", "\%"),
  matrix_sep = " "
)
}
\description{
Save long format mtx data.frame as mtx
}
