% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Plot alluvial from seurat data}
\usage{
plot_alluvial(group1, group2)
}
\arguments{
\item{group1}{vector to be included as group1 (x-axis bar 1)}

\item{group2}{vector to be included as group2 (x-axis bar 2)}

\item{Seurat}{object}
}
\description{
Plot alluvial from seurat data
}
