% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_rank.R
\name{plot_barcode_rank}
\alias{plot_barcode_rank}
\title{Barcode-Rank plot}
\usage{
plot_barcode_rank(data, lower = 10, show.inflection = TRUE, show.knee = FALSE)
}
\value{
ggplot object
}
\description{
Barcode-Rank plot
}
