% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connected_barplot.R
\name{plot_connected_barplot}
\alias{plot_connected_barplot}
\title{Plot connected barplot}
\usage{
plot_connected_barplot(
  population,
  group,
  connected = TRUE,
  bar_width = 0.4,
  y_value = "percent",
  order = FALSE,
  colors = c(),
  label = FALSE
)
}
\arguments{
\item{population}{vector containing population/cluster assignment (y-axis)}

\item{group}{vector containing group or sample assignment (x-axis)}

\item{connected}{Should bars from same group be connected?}

\item{bar_width}{Relative width of bars (integer between 0 and 1)}

\item{y_value}{what statistic to plot ("percent" or "count")}

\item{order}{should populations be ordered by total count (across groups)}

\item{colors}{(named) vector of colors for populations}
}
\value{
ggplot object
}
\description{
Plot connected barplot
}
