% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot dimensional reduction from SCE}
\usage{
plot_dimred(
  object,
  colour_by,
  by_exprs_values = "logcounts",
  max.cutoff = NA,
  min.cutoff = NA,
  order = FALSE,
  decreasing = FALSE,
  na.last = FALSE,
  ...
)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{colour_by}{Variable to use for coloring}

\item{by_exprs_values}{SCE Assay to use}

\item{max.cutoff}{Cutoff value for upper limit (can be integer or 'q' value for quantile: 'q95' sets at 95% quantile)}

\item{min.cutoff}{Cutoff value for lower limit (see max.cutoff)}

\item{order}{Should points plotted in order by their value}

\item{decreasing}{Should ordering be decreasing}

\item{na.last}{Should NA values be last (on top)?}

\item{...}{Passed on to scater::plotReducedDim}
}
\value{
ggplot2 object
}
\description{
Plot dimensional reduction from SCE
}
