% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dots.R
\name{plot_dots}
\alias{plot_dots}
\title{Plot dots}
\usage{
plot_dots(
  sce,
  rows = SingleCellExperiment::colLabels(sce),
  rows_colors = NA,
  rows_colors_width = 0.02,
  rows_distribution_by = NA,
  rows_distribution_fill = NULL,
  rows_distribution_width = 0.05,
  colnames_fun = function(x) {
     return(as.factor(x))
 },
  coldata = SingleCellExperiment::colData(sce),
  assay_color = "logcounts",
  assay_positive = "fg_prob",
  threshold_positive = 0.95,
  threshold_color = 2,
  threshold_fraction = 0.05,
  feature_annotation = NULL,
  feature_annotation_height = 0.1,
  max.cutoff = NA,
  gaps_y = NA,
  scale_fill = scale_fill_viridis_c(option = "turbo")
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing expression matrices}

\item{rows}{colData column defining clustering to be used as rows}

\item{rows_colors}{rows can be color annotated by providing a named color palette here}

\item{rows_colors_width}{Relative (to 1) width of rows color annotation}

\item{rows_distribution_by}{Plot how cluster is distributed among another colData column}

\item{rows_distribution_fill}{Colors palette (color vector) for cluster distribution}

\item{rows_distribution_width}{Relative (to 1) width of rows distribution plot}

\item{colnames_fun}{Function to clean up rownames (such as trimming prefix)}

\item{coldata}{DataFrame containing cluster information for each column in the provided matrices. Defaults to colData(sce)}

\item{assay_color}{Name of SCE assay to use for coloring dots}

\item{assay_positive}{Name of SCE assay to use for gating positive cells}

\item{threshold_positive}{Threshold for gating positive cells}

\item{threshold_color}{Only positive populations having a mean expression above this threshold will be displayed}

\item{threshold_fraction}{Minimum fraction of cells to be included in the plot}

\item{feature_annotation}{Annotation data.frame for features (columns) to be shown as a heatmap}

\item{feature_annotation_height}{Relative (to 1) height of feature annotation heatmap}

\item{max.cutoff}{Cutoff value for upper limit (can be integer or 'q' value for quantile: 'q95' sets at 95\% percentile)}

\item{gaps_y}{Not currently used}

\item{scale_fill}{ggplot scale_fill_continuous function for coloring dots}
}
\value{
ggplot2 combined with patchwork
}
\description{
Plots population/cluster statistics as percentage positive (dot size) and mean expression within positive population (color)
}
