% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_rank.R
\name{plot_feature_rank}
\alias{plot_feature_rank}
\title{Feature Rank plot}
\usage{
plot_feature_rank(
  value,
  group = NULL,
  group_names = NA,
  split = NULL,
  celltype_group = NULL,
  cell_color_by = NULL,
  combine = TRUE,
  nrow = 1,
  threshold_y = NA,
  title = NA,
  remove_ylabel = FALSE,
  remove_ylabel_width = 0.92,
  labels = c(),
  widths = NULL,
  ...
)
}
\arguments{
\item{value}{Value used for ranking and x-axis (i.e. UMI count).}

\item{group}{Grouping variable for comparison (i.e. titration).}

\item{group_names}{Rename group levels (i.e. can be used to append concentration or other info that is marker specific). Named vector (or list of such if split is set).}

\item{split}{Splitting variable (generates individual plot for each group)}

\item{celltype_group}{Celltype for grouping cells (i.e. lineage).}

\item{cell_color_by}{Variable for coloring cells (i.e. celltype). Used for "barcode" plot.}

\item{combine}{Should plot components be combined (default). If FALSE, a list of plot components will be returned.}

\item{nrow}{Number of rows to use when wrapping}

\item{threshold_y}{If not NULL, a threshold line will be drawn at a the set rank. If a single value, same threshold is set for all splits. When splitting, multiple values can be given in a vector (named by split)}

\item{title}{If not NA, title will be set. If a single value title only first panel will get title. When splitting, multiple values can be given in a vector (named by split)}

\item{remove_ylabel}{Should y-labels for all but the first panel be removed?}

\item{remove_ylabel_width}{Decrease relative width of non-first plots by this factor to account for removed y-labels in others}

\item{...}{Passed on to plot_feature_rank_single().}
}
\value{
ggplot2 object or list of ggplot2 objects (if combine=FALSE)
}
\description{
Feature Rank plot
}
