% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_rank.R
\name{plot_feature_rank_combine}
\alias{plot_feature_rank_combine}
\title{Combine Feature Rank plot}
\usage{
plot_feature_rank_combine(
  components,
  title = NA,
  subtitle = NULL,
  height_rank = 8,
  height_density = 2,
  height_barplot = 1.6,
  width_rank = 8,
  width_violin = 2,
  hjust = 0.5,
  vjust = 1,
  label_x = 0.55,
  ...
)
}
\arguments{
\item{components}{List of components (returned from feature_rank_plot() when combine=FALSE)}

\item{height_rank}{Relative height of rank plot}

\item{height_density}{Relative height of density plot}

\item{width_rank}{Relative width of rank plot}

\item{width_violin}{Relative width of violin plot}

\item{hjust}{Horizontal justification of title. Passed on to cowplot::plot_grid()}

\item{vjust}{Vertical justification of title. Passed on to cowplot::plot_grid()}

\item{label_x}{Label x-position. Passed on to cowplot::plot_grid()}

\item{...}{Passed on to cowplot::plot_grid()}
}
\description{
Function for combining components of the feature rank plot.
}
