% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_rank.R
\name{plot_feature_rank_single}
\alias{plot_feature_rank_single}
\title{Feature Rank plot}
\usage{
plot_feature_rank_single(
  value,
  group = NULL,
  group_color = NULL,
  group_names = NA,
  celltype_group = NULL,
  celltype_group_color = NULL,
  cell_color_by = NULL,
  cell_color = NULL,
  barcode_subsample = 200,
  barplot_group = NULL,
  barplot_stack = NULL,
  barplot_stack_color = NULL,
  barplot_show_total = TRUE,
  barplot_total_fontface = "bold",
  barplot_show_ymax = TRUE,
  barplot_alpha = 1,
  show_density = TRUE,
  show_barplot = TRUE,
  show_barcode = TRUE,
  show_violin = TRUE,
  show_squares = TRUE,
  combine = TRUE,
  ties_method = "first",
  rank_normalize = TRUE,
  threshold_y = NA,
  threshold_color = alpha("black", 0.5),
  threshold_size = 1,
  threshold_linetype = "dashed",
  threshold_show_celltype_pct = TRUE,
  threshold_celltype_pct_fontface = "bold",
  threshold_celltype_pct_size = (theme_get()$text$size/(1/0.352777778)),
  labels = c(),
  ...
)
}
\arguments{
\item{value}{Value used for ranking and x-axis (i.e. UMI count).}

\item{group}{Grouping variable for comparison (i.e. titration).}

\item{group_color}{Named vector for coloring groups. Used in density plot.}

\item{group_names}{Rename group levels (i.e. can be used to append concentration or other info that is marker specific)}

\item{celltype_group}{Celltype for grouping cells (i.e. lineage).}

\item{celltype_group_color}{Named vector for coloring celltype_group. Used in violin plot fill.}

\item{cell_color_by}{Variable for coloring cells (i.e. celltype). Used in "barcode" plot.}

\item{cell_color}{Named vector for coloring cells. "Barcode" colors.}

\item{barcode_subsample}{Subsample up to this number of cells from each celltype_group to display in barcode plot}

\item{barplot_group}{Grouping variable for barplots (usually the same as group or celltype_group)}

\item{barplot_stack}{Subgrouping variable for "stacks" within barplot (usually the same as cell_color_by or celltype_group)}

\item{barplot_stack_color}{Named vector for coloring barplot_stack levels}

\item{barplot_show_total}{Should total sums be displayed within barplot? (most meaningful for non-normalized counts)}

\item{barplot_total_fontface}{Font face for total in barplot. Default "bold".}

\item{barplot_show_ymax}{Should maximum y-value tick be shown for barplot? (most meaningful for non-normalized counts)}

\item{barplot_alpha}{Barplot alpha (can be reduced to make total count more readable)}

\item{show_density}{Should density plot be drawn?}

\item{show_barplot}{Should barplot be drawn?}

\item{show_barcode}{Should barcode plot be drawn?}

\item{show_violin}{Should violin plot be drawn?}

\item{show_squares}{Should group squares above barplot be drawn?}

\item{combine}{Should plot components be combined (default). If FALSE, a list of plot components will be returned.}

\item{ties_method}{ties.method passed on to rank().}

\item{threshold_y}{If not NULL, a threshold line will be drawn at a the set rank}

\item{threshold_color}{Threshold line color}

\item{threshold_size}{Threshold line size}

\item{threshold_linetype}{Threshold line type}

\item{threshold_show_celltype_pct}{Show percent above threshold in celltype component}

\item{threshold_celltype_pct_fontface}{Label face for percent above threshold in celltype component}

\item{threshold_celltype_pct_size}{Label size for percent above threshold in celltype component}

\item{labels}{Named vector of axis labels (set element to NULL to remove axis title). Default: c("x.rank"="Value", "y.rank"="Rank", "y.density"="Density", "y.barplot"="Sum")}

\item{...}{Passed on to plot_feature_rank_combine(). Useful for setting relative component sizes.}

\item{barplot_ymax_size}{Font size for ymax text}

\item{rank_normalized}{Should ranks be normalized by group (TRUE) or show as absolute counts (FALSE). Default TRUE.}
}
\value{
ggplot2 object or list of ggplot2 objects (if combine=FALSE)
}
\description{
Feature Rank plot
}
