% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quadrants.R
\name{plot_quadrants}
\alias{plot_quadrants}
\title{Plot quadrants}
\usage{
plot_quadrants(
  plotData,
  x,
  y,
  color,
  wrap = NULL,
  trans = "log1p",
  breaks = c(0, 1, 2, 3, 5, 10, 25, 50, 100, 200, 500, 1000),
  threshold_x = 0.5,
  threshold_y = 0.5,
  text_pos_lower = log1p(-0.25),
  jitter = position_jitter(width = 0.18, height = 0.18, seed = 124)
)
}
\arguments{
\item{plotData}{data frame containing data to be plotted}

\item{x}{column name for x data}

\item{y}{column name for y data}

\item{color}{column name for colouring}

\item{wrap}{column name for wrapping plot}

\item{trans}{transformation function for x and y scales (log1p default)}

\item{breaks}{breaks for x and y scales}

\item{threshold_x}{x gating threshold}

\item{threshold_y}{y gating threshold}

\item{text_pos_lower}{text position}

\item{jitter}{jitter function for geom_points}
}
\value{
ggplot
}
\description{
Plot quadrants
}
