% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_load.R
\name{read_kallisto_data}
\alias{read_kallisto_data}
\title{Load and reformat kallisto output for loading}
\usage{
read_kallisto_data(path, name = "cells_x_genes")
}
\description{
Load and reformat kallisto output for loading
}
