% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVI.R
\name{scVI_load}
\alias{scVI_load}
\title{Load scVI output from adata object into SingleCellExperiment}
\usage{
scVI_load(
  adata,
  sce = NULL,
  obs_ignore = c("_scvi_labels", "_scvi_batch"),
  obs_include = "all",
  obs_overwrite = FALSE,
  obs_prefix = "",
  barcodes = intersect,
  reducedDim_prefix = "",
  reducedDim_include = c("X_scVI", "X_umap"),
  reducedDimNames_replace = "^X_",
  RNA_layer = "scvi_normalized",
  altexp_name = "VI_RNA",
  rownames_prefix = "VI_"
)
}
\arguments{
\item{adata}{AnnData object containing scVI output}

\item{sce}{Existing sce object. If supplied, coldata is added to this object and scVI normalized expression is added as an altExp}

\item{obs_ignore}{Which columns in adata$obs should be ignored before importing?}

\item{obs_include}{Which columns in adata$obs should be included (default "all")}

\item{obs_overwrite}{Should columns existing in the sce be overwritten?}

\item{obs_prefix}{Prefix obs columns before importing}

\item{barcodes}{Function to handle if different barcodes are included in either SCE or adata objects? Intersect to include only barcodes present in both objects, union to include all barcodes - will fill missing values with NA and missing expression with 0 (default: intersect)}

\item{reducedDim_prefix}{Prefix reducedDim names before adding to sce}

\item{reducedDim_include}{Which reducedDims should be transferred from adata$obsm}

\item{reducedDimNames_replace}{String to replace in obsm names before adding to the sce (removes "^X_" by default)}

\item{RNA_layer}{Name of adata$layers that contains the scVI normalized expression matrix}

\item{altexp_name}{Name of altExp to place scVI normalized SCE object if existing sce is supplied}

\item{rownames_prefix}{Prefix for rownames in scVI normalized SCE}
}
\value{
Returns a SingleCellExperiment (SCE). If existing sce is not given the returned SCE contains scVI normalized expression, else scVI normalized expression is added as an altExp
}
\description{
Load scVI output from adata object into SingleCellExperiment
}
