% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cells.R
\name{seurat_annotate_SingleR}
\alias{seurat_annotate_SingleR}
\title{Run Annotate_SingleR on Seurat object}
\usage{
seurat_annotate_SingleR(
  object,
  assay = "RNA",
  slot = "data",
  metadata_column = "SingleR",
  annotation_column = "pruned.labels",
  ...
)
}
\description{
Run Annotate_SingleR on Seurat object
}
