% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_labels.R
\name{seurat_dimred_cluster_labels}
\alias{seurat_dimred_cluster_labels}
\title{Make labels for groups in DimRed plots}
\usage{
seurat_dimred_cluster_labels(
  object,
  reduction = NULL,
  dims = c(1, 2),
  group.by = "ident",
  FUN = median
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Label of reduction object to use (uses the last created reduction by default)}

\item{dims}{Which dimensions should be used?}

\item{group.by}{Meta data column to label clusters}

\item{FUN}{Function called on reduced dimensions to calculate label position}
}
\value{
tibble of labels and their positions
}
\description{
Assign labels to positions in reduced dimensional space. Can be used to label clusters by other meta.data than their coloring.
}
