% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred_encircle.R
\name{seurat_dimred_encircle_add}
\alias{seurat_dimred_encircle_add}
\title{Encircle with Seurat meta.data}
\usage{
seurat_dimred_encircle_add(
  object,
  reduction = NULL,
  group.by = NULL,
  split.by = NULL,
  cells = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Name of reduction in Seurat object. Uses last in list by default.}

\item{group.by}{Name of meta data column used for grouping contours}

\item{split.by}{Name of meta data column included in the ggplot data.frame (for usage in splitting)}

\item{cells}{Subset of cells to be used for calculating KDE contours. Cell barcodes or indices.}
}
\value{
list of ggplot elements
}
\description{
Encircle by Kernel Density Estimate contours
}
