% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_labels.R
\name{seurat_dimred_labels_add}
\alias{seurat_dimred_labels_add}
\title{Add labels to DimPlot with annotation from seurat object}
\usage{
seurat_dimred_labels_add(
  object,
  reduction = NULL,
  dims = c(1, 2),
  group.by = "ident",
  FUN = median,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Label of reduction object to use (uses the last created reduction by default)}

\item{dims}{Which dimensions should be used?}

\item{group.by}{Meta data column to label clusters}

\item{FUN}{Function called on reduced dimensions to calculate label position}

\item{...}{Passed on to scutility::dimred_labels_add}
}
\value{
list of ggplot2 functions to be added to an existing plot
}
\description{
Wrapper function for calling seurat_dimred_cluster_labels and dimred_labels_add
}
