% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{seurat_plot_alluvial}
\alias{seurat_plot_alluvial}
\title{Plot alluvial from seurat data}
\usage{
seurat_plot_alluvial(object, group1, group2)
}
\arguments{
\item{object}{Seurat object}

\item{group1}{columns to be included as group1 (x-axis bar 1)}

\item{group2}{columns to be included as group2 (x-axis bar 2)}
}
\description{
Plot alluvial from seurat data
}
