% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connected_barplot.R
\name{seurat_plot_connected_barplot}
\alias{seurat_plot_connected_barplot}
\title{Plot connected barplot from seurat object}
\usage{
seurat_plot_connected_barplot(object, group.by = "ident", split.by, ...)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Meta data column containing population/cluster assignment (y-axis)}

\item{split.by}{Meta data column containing group or sample assignment (x-axis)}
}
\value{
ggplot object
}
\description{
Plot connected barplot from seurat object
}
