% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred_counts.R
\name{seurat_plot_dimred_counts}
\alias{seurat_plot_dimred_counts}
\title{DimensionReduction with counts plot}
\usage{
seurat_plot_dimred_counts(
  object,
  feature,
  split.by = NULL,
  colors = c("#000033", "#3377FF", "#33AAFF", "#33CC33", "orange", "red"),
  limits = c(0, NA),
  order = FALSE,
  slot = "counts",
  max.cutoff = "q95",
  pt.size = 0.75,
  legend.key.width = unit(5, "mm"),
  legend.key.height = unit(2, "mm"),
  legend.text = element_text(angle = 45),
  combine = TRUE,
  ...
)
}
\arguments{
\item{split.by}{Meta.data column in seurat object for splitting}

\item{colors}{Color palette}

\item{limits}{Limits for color palette (for counts it makes sense that lower limit is 0; DEFAULT: c(0, NA))}

\item{order}{Should cells be ordered by expression}

\item{combine}{Should plots be combined or returned as a list of plots?}

\item{...}{Passed on to Seurat::FeaturePlot}
}
\value{
ggplot2 object(s)
}
\description{
Very crude and hardcoded function that calls Seurat::FeaturePlot. Needs a remake for more flexibility.
}
