% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_rank.R
\name{seurat_plot_feature_rank}
\alias{seurat_plot_feature_rank}
\title{Feature Rank plot Seurat}
\usage{
seurat_plot_feature_rank(
  object,
  feature,
  group.by = NULL,
  split.by = NULL,
  celltype_group = NULL,
  cell_color_by = NULL,
  slot = "counts",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Meta.data column for ranking and x-axis (i.e. UMI count).}

\item{group.by}{Meta.data column for grouping variable for comparison (i.e. titration).}

\item{split.by}{Meta.data column for splitting plots}

\item{celltype_group}{Meta.data column celltype for grouping cells (i.e. lineage).}

\item{cell_color_by}{Meta.data column for coloring cells (i.e. celltype). Used for "barcode" plot.}

\item{slot}{Slot that the value should be extracted from (default "counts")}
}
\value{
ggplot2 object or list of ggplot2 objects (if combine=FALSE)
}
\description{
Create feature rank plot from Seurat object
}
