% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clusters.R
\name{subcluster_number}
\alias{subcluster_number}
\title{Number subclusters by parent cluster}
\usage{
subcluster_number(subcluster, cluster)
}
\arguments{
\item{subcluster}{vector of subcluster assignments (one value for each cell)}

\item{cluster}{vector of parent cluster assignment (one value for each cell)}
}
\value{
vector of cluster annotations
}
\description{
Name subclusters by their parent cluster and numbering. The subcluster with the highest number of cells will be number 1 and so on.
}
