% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subset_matrix}
\alias{subset_matrix}
\title{Subset a matrix}
\usage{
subset_matrix(data, barcodes, includeAll = TRUE, na.value = 0)
}
\value{
A subsetted sparse Matrix
}
\description{
If includeAll is TRUE (default) and if a barcode is not present in the
matrix, set its values to 0 to allow incorporation into a common object
}
