% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{translate_ensg_to_symbol}
\alias{translate_ensg_to_symbol}
\title{Translate ENSG annotation to gene symbol}
\usage{
translate_ensg_to_symbol(raw_mtx, t2g.file)
}
\value{
Count Matrix with translated rownames
}
\description{
Translate ENSG annotation to gene symbol
}
