% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_load.R
\name{modalities_load_cellranger_count}
\alias{modalities_load_cellranger_count}
\title{Load Cell Ranger data and divide by modality}
\usage{
modalities_load_cellranger_count(
  path,
  modalities = c("RNA", "ADT", "HTO"),
  folder = "raw_feature_bc_matrix",
  hto.pattern = "^hto",
  gex.listname = "Gene Expression",
  adt.listname = "Antibody Capture"
)
}
\value{
list of modality count matrices
}
\description{
Load Cell Ranger data and divide by modality
}
