% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_load.R
\name{modalities_load_kallisto}
\alias{modalities_load_kallisto}
\title{Load Kallisto data by modality}
\usage{
modalities_load_kallisto(paths, modalities = c("GEX", "ADT", "HTO"))
}
\value{
list of modality count matrices
}
\description{
Load Kallisto data by modality
}
