% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_rank.R
\name{plot_barcode_rank}
\alias{plot_barcode_rank}
\title{Barcode-Rank plot}
\usage{
plot_barcode_rank(data, lower = 10, show.inflection = TRUE, show.knee = FALSE)
}
\arguments{
\item{data}{Count matrix}

\item{lower}{Lower threshold passed on to DropletUtils::barcodeRanks()}

\item{show.inflection}{Plot line at inflection point?}

\item{show.knee}{Plot line at knee point?}
}
\value{
ggplot object
}
\description{
Barcode-Rank plot
}
