% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connected_barplot.R
\name{plot_connected_barplot}
\alias{plot_connected_barplot}
\title{Plot connected barplot}
\usage{
plot_connected_barplot(
  population,
  group,
  y_value = "percent",
  order = FALSE,
  colors = c(),
  label = FALSE
)
}
\arguments{
\item{population}{vector containing population/cluster assignment (y-axis)}

\item{group}{vector containing group or sample assignment (x-axis)}

\item{order}{should populations be ordered by total count (across groups)}

\item{value}{what statistic to plot ("percent" or "count")}

\item{color}{(named) vector of colors for populations}
}
\value{
ggplot object
}
\description{
Plot connected barplot
}
