% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimred.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot dimensional reduction from SCE}
\usage{
plot_dimred(
  object,
  colour_by,
  colour_list_in_metadata = "colors",
  colours_reduce_labels = TRUE,
  features_add = c(),
  dimred = "UMAP",
  columns = NULL,
  by_exprs_values = "logcounts",
  seed = 12232,
  order = FALSE,
  order_function = function(x) {
     order(x, decreasing = FALSE, na.last = FALSE)
 },
  shuffle = FALSE,
  point_size = 0.5,
  point_alpha = 1,
  max.cutoff = NA,
  min.cutoff = NA,
  scale_color = NULL,
  value_convert = NULL,
  text_by = NULL,
  text_by_colour = "black",
  text_by_dimred_summary_fun = median,
  text_by_size = NA,
  text_by_fontface = "bold",
  text_by_padding = unit(0.01, units = "npc"),
  text_by_fill = alpha(c("white"), 0.75),
  text_repel = TRUE,
  text_repel_max_overlaps = Inf,
  coldata_exclude_class = c("CompressedSplitDFrameList"),
  rasterise = FALSE,
  rasterise_dev = "cairo",
  rasterise_dpi = 300,
  rasterise_scale = 1,
  ...
)
}
\arguments{
\item{object}{SingleCellExperiment object}

\item{colour_by}{Variable to use for coloring}

\item{colour_list_in_metadata}{If set, function will search for a manual colouring vector in the metadata(sce)[[colour_list_in_metadata]][[colour_by]]}

\item{colours_reduce_labels}{If TRUE, reduce colour legend to only show values included in the plot}

\item{features_add}{Fetch additional features to data.frame given to ggplot (for specialized plotting)}

\item{dimred}{Which dimensional reductions should be used?}

\item{columns}{Vector of columns in SCE to plot}

\item{by_exprs_values}{SCE Assay to use}

\item{seed}{Seed for random operations}

\item{order}{Should points plotted in order by their value}

\item{order_function}{What function should be used for ordering?}

\item{shuffle}{Should points be plotted in random order}

\item{point_size}{Size of individual points}

\item{point_alpha}{Alpha level of each point}

\item{max.cutoff}{Cutoff value for upper limit (can be integer or 'q' value for quantile: 'q95' sets at 95\% percentile)}

\item{min.cutoff}{Cutoff value for lower limit (see max.cutoff)}

\item{scale_color}{Color scale to use (viridis by default)}

\item{value_convert}{Function to convert values before plotting (If NULL, no conversion)}

\item{text_by}{Variable to use for labelling}

\item{text_by_colour}{Label color if text_by is set (black by default)}

\item{text_by_dimred_summary_fun}{Which function should be used to place text_by (default median)}

\item{text_by_size}{Text size if text_by is set}

\item{text_by_fontface}{Fontface is text_by is set (default "bold")}

\item{text_by_padding}{Label padding if text_by is set.}

\item{text_by_fill}{Label fill if text_by is set (alpha("white", 0.75) default)}

\item{text_repel}{Should text labels be repelled from each other?}

\item{text_repel_max_overlaps}{Max overlaps for ggrepel}

\item{coldata_exclude_class}{By default loads all colData except the columns of the classes included in this vector}

\item{rasterise}{Should points be rasterised (ggrastr)?}

\item{rasterise_dev}{What device should be used for rasterisation?}

\item{rasterise_dpi}{What DPI should be used for rasterisation?}

\item{rasterise_scale}{What scale should be used for rasterisation?}

\item{...}{Passed on to makePerCellDF}
}
\value{
ggplot2 object
}
\description{
Plot dimensional reduction from SCE
}
